requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
asgi-lifespan==2.1.0
async-asgi-testclient==1.4.11
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
black==24.2.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.0
MarkupSafe==2.1.4
multidict==6.0.5
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycares==4.4.0
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pytest==8.0.2
pytest-aiohttp==1.0.5
pytest-asyncio==0.23.5
pytest-dotenv==0.5.2
pytest-env==1.1.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
types-passlib==1.7.7.20240106
types-pyasn1==0.5.0.20240301
types-python-jose==3.3.4.20240106
types-requests==2.31.0.20240125
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.26.0
yarl==1.9.4
